refactor(compiler): collapse generate-init.ts to a ~140-line orchestrator (#1021)
Stage D of issue #1021 — analysis-on-IR refactor. Final structural
pass: `generateInitFunction` reads like a pipeline again.

Before: 450 lines mixing preamble emission, analysis, scope routing,
declaration collection, DOM-ref generation, reactive emission, string
post-processing, module-level emission, and import resolution — all
inline.

After: 144 lines. Each non-trivial stage lives in its own file:

  packages/jsx/src/ir-to-client-js/
    element-refs.ts         (89 lines) — $/$t/$c declarations
    child-components.ts    (101 lines) — @bf-child import markers +
                                         IR-tree name collection
    init-declarations.ts   (183 lines) — classifyLocalDeclarations
                                         + emitSortedDeclarations
    emit-module-level.ts    (83 lines) — module-level `var X = X ?? …`
                                         emission + final import line
                                         resolution

The orchestrator now reads as:

  preamble → analysis (graph + scope + prop usage)
          → init body emission
          → hydrate line
          → props rename (pure-function helper, still in-file)
          → placeholder replacement

File sizes:

  generate-init.ts:  450 → 144 lines (-306)
  new split files:   +456 lines
  net:               +150 lines, but each file is focused

Invariant checked:
- All 326 compiled `.client.js` files under `site/ui/dist/components/`
  are byte-identical to `main`.
- jsx unit (763), adapter-tests (214), hono (80), go-template (63),
  client (212), form (40), ui (1157) — all green (2529 tests).

Meets the Stage A spec target: file length under 150 lines, pipeline
structure visible, each emitter is a pure function of typed inputs.

Not in this PR (tracked as Stage E / follow-up):
- Props-rename regex hack still lives in `generate-init.ts` as a
  private helper (`renamePropsObjectInInitBody`). Removing it needs
  analyzer-time pre-rewriting of every IR string field that can carry
  a prop reference (`SignalInfo.initialValue`, `MemoInfo.computation`,
  `FunctionInfo.body`, …), which is a substantial analyzer change and
  was deemed high-risk for the Stage D byte-identical invariant.
